Grabbed object orientation and animation triggers
Grabbed objects should now always be in the grip of the tentacles, due
to a change in the objects position (indicated in PickupObject.cs).
Animations are now dependent on states that the player is in rather than
just clicks. If the player's acc(located in simple_movement.cs) is
greater than 0, the animation for swimming will play. If the carrying
bool (located in PickupObject.cs) is true, the grabbing animation will
play. When carrying bool is not true, the grabbing animation will stop