:mount_fuji:  :leaves: :lantern:
So messed with the terrain so its a bit hilly. Imported Gina's model for
the top shelf but its not quite a perfect fit. Also made some changes to
the map as far as where the gauntlet will go. Added colliders to the
trees

Things to fix: add rocks to bottom half, add the top shelf. fix caustics
and lighting. yee